refactor: relocate Kitchen, Dishes, and Ingredients views into a new Kitchen subdirectory and update router paths.
