+++ src/views/Kitchen/Ingredients.vue
@@ -44,9 +44,9 @@ import { ref, onMounted } from "vue";
 import axios from "axios";
 import { message } from "ant-design-vue";
 import { PlusOutlined, DeleteOutlined } from "@ant-design/icons-vue";
-import { useAuthStore } from '../stores/auth';
-import AddIngredientDrawer from '../components/AddIngredientDrawer.vue';
-import placeholderImage from '../assets/cocktail_placeholder.png';
+import { useAuthStore } from '../../stores/auth';
+import AddIngredientDrawer from '../../components/AddIngredientDrawer.vue';
+import placeholderImage from '../../assets/cocktail_placeholder.png';
 
 const ingredients = ref([]);
 const showAddDrawer = ref(false);
